Get rid of orphan instances.
